refactor(types): remove unused types, streamline player functionality
