+++ src/modules/track/composables/useTrackSelection.ts
@@ -1,9 +1,6 @@
 import { ref, computed, type Ref, type ComputedRef } from 'vue';
 import type { Track } from '../types';
 
-/**
- * Composable for managing track selection for bulk operations
- */
 export function useTrackSelection() {
   // === STATE ===
   const selectedTracks = ref<string[]>([]);
@@ -16,9 +13,6 @@ export function useTrackSelection() {
   const selectedCount: ComputedRef<number> = computed(() => selectedTracks.value.length);
 
   // === METHODS ===
-  /**
-   * Toggle selection of a track
-   */
   function toggleTrackSelection(id: string): void {
     const index = selectedTracks.value.indexOf(id);
     if (index === -1) {
@@ -28,45 +22,27 @@ export function useTrackSelection() {
     }
   }
 
-  /**
-   * Check if a track is selected
-   */
   function isTrackSelected(id: string): boolean {
     return selectedTracks.value.includes(id);
   }
 
-  /**
-   * Select all tracks
-   */
   function selectAllTracks(tracks: Track[]): void {
     selectedTracks.value = tracks.map(track => track.id);
   }
 
-  /**
-   * Clear all selected tracks
-   */
   function clearSelectedTracks(): void {
     selectedTracks.value = [];
   }
 
-  /**
-   * Select multiple tracks by ID
-   */
   function selectTracks(trackIds: string[]): void {
     selectedTracks.value = [...trackIds];
   }
 
-  /**
-   * Add tracks to selection
-   */
   function addToSelection(trackIds: string[]): void {
     const newIds = trackIds.filter(id => !selectedTracks.value.includes(id));
     selectedTracks.value.push(...newIds);
   }
 
-  /**
-   * Remove tracks from selection
-   */
   function removeFromSelection(trackIds: string[]): void {
     selectedTracks.value = selectedTracks.value.filter(id => !trackIds.includes(id));
   }
